✨  Feat: 유저 종합점수 계산 로직 추가

+++ src/main/java/cmc/peerna/web/dto/responseDto/MemberResponseDto.java
@@ -39,7 +39,7 @@ public class MemberResponseDto {
         private Part part;
         private Job job;
 
-        private Long totalScore;
+        private Integer totalScore;
         private String oneLiner;
     }
